requirements.txt
django==1.11.5
django-debug-toolbar
uwsgi
